Move the message types to a separate file
